chore: restructure code to support multiple DID Document formats
