+++ packages/kilt-did/src/DidDetails/FullDidDetails.spec.ts
@@ -11,7 +11,6 @@
 
 import { KeyRelationship } from '@kiltprotocol/types'
 import { BN } from '@polkadot/util'
-import { base58Encode } from '@polkadot/util-crypto'
 import type { IDidKeyDetails } from '@kiltprotocol/types'
 import { mapCallToKeyRelationship } from './FullDidDetails.utils'
 import { FullDidDetails, FullDidDetailsCreationOpts } from './FullDidDetails'
@@ -214,166 +213,6 @@ describe('functional tests', () => {
         "${did}#4",
       ]
     `)
-  })
-
-  it('exports the expected application/json W3C DID Document with an Ed25519 authentication key, two x25519 encryption keys, an Sr25519 assertion key, an Ecdsa delegation key, and some service endpoints', () => {
-    const ecdsaKey: IDidKeyDetails = {
-      id: `${did}#5`,
-      controller: did,
-      includedAt: 200,
-      type: 'ecdsa',
-      publicKeyHex:
-        '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
-    }
-    const fullDidDetails = new FullDidDetails({
-      ...didDetails,
-      keyRelationships: {
-        [KeyRelationship.authentication]: [keys[0].id],
-        [KeyRelationship.keyAgreement]: [keys[1].id, keys[2].id],
-        [KeyRelationship.assertionMethod]: [keys[3].id],
-        [KeyRelationship.capabilityDelegation]: [ecdsaKey.id],
-      },
-      keys: keys.concat([ecdsaKey]),
-    })
-
-    const didDoc = fullDidDetails.toDidDocument('application/json')
-
-    expect(didDoc.id).toMatch(fullDidDetails.did)
-
-    expect(didDoc.authentication).toHaveLength(1)
-    expect(didDoc.authentication).toContainEqual(`${fullDidDetails.did}#1`)
-
-    expect(didDoc.keyAgreement).toHaveLength(2)
-    expect(didDoc.keyAgreement).toContainEqual(`${fullDidDetails.did}#2`)
-    expect(didDoc.keyAgreement).toContainEqual(`${fullDidDetails.did}#3`)
-
-    expect(didDoc.assertionMethod).toHaveLength(1)
-    expect(didDoc.assertionMethod).toContainEqual(`${fullDidDetails.did}#4`)
-
-    expect(didDoc.capabilityDelegation).toHaveLength(1)
-    expect(didDoc.capabilityDelegation).toContainEqual(
-      `${fullDidDetails.did}#5`
-    )
-
-    expect(didDoc.verificationMethod).toHaveLength(5)
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#1`,
-      controller: fullDidDetails.did,
-      type: 'Ed25519VerificationKey2018',
-      publicKeyBase58: base58Encode(keys[0].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#2`,
-      controller: fullDidDetails.did,
-      type: 'X25519KeyAgreementKey2019',
-      publicKeyBase58: base58Encode(keys[1].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#3`,
-      controller: fullDidDetails.did,
-      type: 'X25519KeyAgreementKey2019',
-      publicKeyBase58: base58Encode(keys[2].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#4`,
-      controller: fullDidDetails.did,
-      type: 'Sr25519VerificationKey2020',
-      publicKeyBase58: base58Encode(keys[3].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#5`,
-      controller: fullDidDetails.did,
-      type: 'EcdsaSecp256k1VerificationKey2019',
-      publicKeyBase58: base58Encode(ecdsaKey.publicKeyHex),
-    })
-
-    expect(didDoc.service).toHaveLength(2)
-    expect(didDoc.service).toContainEqual(services[0])
-    expect(didDoc.service).toContainEqual(services[1])
-  })
-
-  it('exports the expected application/json+ld W3C DID Document with only an authentication key', () => {
-    const ecdsaKey: IDidKeyDetails = {
-      id: `${did}#5`,
-      controller: did,
-      includedAt: 200,
-      type: 'ecdsa',
-      publicKeyHex:
-        '0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee',
-    }
-    const fullDidDetails = new FullDidDetails({
-      ...didDetails,
-      keyRelationships: {
-        [KeyRelationship.authentication]: [keys[0].id],
-        [KeyRelationship.keyAgreement]: [keys[1].id, keys[2].id],
-        [KeyRelationship.assertionMethod]: [keys[3].id],
-        [KeyRelationship.capabilityDelegation]: [ecdsaKey.id],
-      },
-      keys: keys.concat([ecdsaKey]),
-    })
-
-    const didDoc = fullDidDetails.toDidDocument('application/json+ld')
-
-    expect(didDoc.id).toMatch(fullDidDetails.did)
-
-    expect(didDoc['@context']).toHaveLength(1)
-    expect(didDoc['@context']).toContainEqual('https://www.w3.org/ns/did/v1')
-
-    expect(didDoc.authentication).toHaveLength(1)
-    expect(didDoc.authentication).toContainEqual(`${fullDidDetails.did}#1`)
-
-    expect(didDoc.keyAgreement).toHaveLength(2)
-    expect(didDoc.keyAgreement).toContainEqual(`${fullDidDetails.did}#2`)
-    expect(didDoc.keyAgreement).toContainEqual(`${fullDidDetails.did}#3`)
-
-    expect(didDoc.assertionMethod).toHaveLength(1)
-    expect(didDoc.assertionMethod).toContainEqual(`${fullDidDetails.did}#4`)
-
-    expect(didDoc.capabilityDelegation).toHaveLength(1)
-    expect(didDoc.capabilityDelegation).toContainEqual(
-      `${fullDidDetails.did}#5`
-    )
-
-    expect(didDoc.verificationMethod).toHaveLength(5)
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#1`,
-      controller: fullDidDetails.did,
-      type: 'Ed25519VerificationKey2018',
-      publicKeyBase58: base58Encode(keys[0].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#2`,
-      controller: fullDidDetails.did,
-      type: 'X25519KeyAgreementKey2019',
-      publicKeyBase58: base58Encode(keys[1].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#3`,
-      controller: fullDidDetails.did,
-      type: 'X25519KeyAgreementKey2019',
-      publicKeyBase58: base58Encode(keys[2].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#4`,
-      controller: fullDidDetails.did,
-      type: 'Sr25519VerificationKey2020',
-      publicKeyBase58: base58Encode(keys[3].publicKeyHex),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${fullDidDetails.did}#5`,
-      controller: fullDidDetails.did,
-      type: 'EcdsaSecp256k1VerificationKey2019',
-      publicKeyBase58: base58Encode(ecdsaKey.publicKeyHex),
-    })
-
-    expect(didDoc.service).toHaveLength(2)
-    expect(didDoc.service).toContainEqual(services[0])
-    expect(didDoc.service).toContainEqual(services[1])
-  })
-
-  it('does not export a DID Document with an unsupported format', () => {
-    const fullDidDetails = new FullDidDetails(didDetails)
-    expect(() => fullDidDetails.toDidDocument('text/html')).toThrow()
   })
 })
