+++ packages/kilt-did/src/DidDetails/LightDidDetails.spec.ts
@@ -11,10 +11,9 @@
  * @group unit/did
  */
 
-import { Crypto } from '@kiltprotocol/utils'
 import type { IServiceDetails } from '@kiltprotocol/types'
 import { hexToU8a } from '@polkadot/util'
-import { base58Encode } from '@polkadot/util-crypto'
+import { encodeAddress } from '@polkadot/util-crypto'
 import { LightDidDetails, LightDidDetailsCreationOpts } from './LightDidDetails'
 import type { INewPublicKey } from '../types'
 
@@ -25,8 +24,8 @@ describe('Light DID v1 tests', () => {
   const encPublicKey = hexToU8a(
     '0xbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb'
   )
-  const address = Crypto.encodeAddress(authPublicKey, 38)
-  let authenticationDidKeyDetails: INewPublicKey = {
+  const address = encodeAddress(authPublicKey, 38)
+  const authenticationDidKeyDetails: INewPublicKey = {
     publicKey: authPublicKey,
     type: 'ed25519',
   }
@@ -113,188 +112,5 @@ describe('Light DID v1 tests', () => {
     const did = new LightDidDetails(didCreationDetails)
     // no concat of : and encoded details
     expect(did.did).toEqual(`did:kilt:light:01${address}`)
-  })
-
-  it('exports the expected application/json W3C DID Document with an Ed25519 authentication key, an x25519 encryption key, and some service endpoints', () => {
-    encryptionDidKeyDetails = {
-      publicKey: encPublicKey,
-      type: 'x25519',
-    }
-    services = [
-      {
-        id: `service1`,
-        type: 'messaging',
-        serviceEndpoint: 'example.com',
-      },
-      {
-        id: `service2`,
-        type: 'telephone',
-        serviceEndpoint: '123344',
-      },
-    ]
-
-    const didCreationDetails: LightDidDetailsCreationOpts = {
-      authenticationKey: authenticationDidKeyDetails,
-      encryptionKey: encryptionDidKeyDetails,
-      services,
-    }
-    const didDetails = new LightDidDetails(didCreationDetails)
-    const didDoc = didDetails.toDidDocument('application/json')
-
-    expect(didDoc.id).toMatch(didDetails.did)
-
-    expect(didDoc.authentication).toHaveLength(1)
-    expect(didDoc.authentication).toContainEqual(
-      `${didDetails.did}#authentication`
-    )
-
-    expect(didDoc.keyAgreement).toHaveLength(1)
-    expect(didDoc.keyAgreement).toContainEqual(`${didDetails.did}#encryption`)
-
-    expect(didDoc.assertionMethod).toBeUndefined()
-
-    expect(didDoc.capabilityDelegation).toBeUndefined()
-
-    expect(didDoc.verificationMethod).toHaveLength(2)
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${didDetails.did}#authentication`,
-      controller: didDetails.did,
-      type: 'Ed25519VerificationKey2018',
-      publicKeyBase58: base58Encode(authenticationDidKeyDetails.publicKey),
-    })
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${didDetails.did}#encryption`,
-      controller: didDetails.did,
-      type: 'X25519KeyAgreementKey2019',
-      publicKeyBase58: base58Encode(encryptionDidKeyDetails.publicKey),
-    })
-
-    expect(didDoc.service).toHaveLength(2)
-    expect(didDoc.service).toContainEqual({
-      id: `${didDetails.did}#service1`,
-      type: 'messaging',
-      serviceEndpoint: 'example.com',
-    })
-    expect(didDoc.service).toContainEqual({
-      id: `${didDetails.did}#service2`,
-      type: 'telephone',
-      serviceEndpoint: '123344',
-    })
-  })
-
-  it('exports the expected application/json W3C DID Document with an Sr25519 authentication key', () => {
-    authenticationDidKeyDetails = {
-      publicKey: authPublicKey,
-      type: 'sr25519',
-    }
-    const didCreationDetails: LightDidDetailsCreationOpts = {
-      authenticationKey: authenticationDidKeyDetails,
-    }
-    const didDetails = new LightDidDetails(didCreationDetails)
-    const didDoc = didDetails.toDidDocument('application/json')
-
-    expect(didDoc.id).toMatch(didDetails.did)
-
-    expect(didDoc.authentication).toHaveLength(1)
-    expect(didDoc.authentication).toContainEqual(
-      `${didDetails.did}#authentication`
-    )
-
-    expect(didDoc.keyAgreement).toBeUndefined()
-
-    expect(didDoc.assertionMethod).toBeUndefined()
-
-    expect(didDoc.capabilityDelegation).toBeUndefined()
-
-    expect(didDoc.verificationMethod).toHaveLength(1)
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${didDetails.did}#authentication`,
-      controller: didDetails.did,
-      type: 'Sr25519VerificationKey2020',
-      publicKeyBase58: base58Encode(authenticationDidKeyDetails.publicKey),
-    })
-
-    expect(didDoc.service).toBeUndefined()
-  })
-
-  it('exports the expected application/json W3C DID Document with an Ecdsa authentication key', () => {
-    authenticationDidKeyDetails = {
-      publicKey: authPublicKey,
-      type: 'ecdsa',
-    }
-    const didCreationDetails: LightDidDetailsCreationOpts = {
-      authenticationKey: authenticationDidKeyDetails,
-    }
-    const didDetails = new LightDidDetails(didCreationDetails)
-    const didDoc = didDetails.toDidDocument('application/json')
-
-    expect(didDoc.id).toMatch(didDetails.did)
-
-    expect(didDoc.authentication).toHaveLength(1)
-    expect(didDoc.authentication).toContainEqual(
-      `${didDetails.did}#authentication`
-    )
-
-    expect(didDoc.keyAgreement).toBeUndefined()
-
-    expect(didDoc.assertionMethod).toBeUndefined()
-
-    expect(didDoc.capabilityDelegation).toBeUndefined()
-
-    expect(didDoc.verificationMethod).toHaveLength(1)
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${didDetails.did}#authentication`,
-      controller: didDetails.did,
-      type: 'EcdsaSecp256k1VerificationKey2019',
-      publicKeyBase58: base58Encode(authenticationDidKeyDetails.publicKey),
-    })
-
-    expect(didDoc.service).toBeUndefined()
-  })
-
-  it('exports the expected application/json+ld W3C DID Document with only an authentication key', () => {
-    authenticationDidKeyDetails = {
-      publicKey: authPublicKey,
-      type: 'sr25519',
-    }
-    const didCreationDetails: LightDidDetailsCreationOpts = {
-      authenticationKey: authenticationDidKeyDetails,
-    }
-    const didDetails = new LightDidDetails(didCreationDetails)
-    const didDoc = didDetails.toDidDocument('application/json+ld')
-
-    expect(didDoc.id).toMatch(didDetails.did)
-
-    expect(didDoc['@context']).toHaveLength(1)
-    expect(didDoc['@context']).toContainEqual('https://www.w3.org/ns/did/v1')
-
-    expect(didDoc.authentication).toHaveLength(1)
-    expect(didDoc.authentication).toContainEqual(
-      `${didDetails.did}#authentication`
-    )
-
-    expect(didDoc.keyAgreement).toBeUndefined()
-
-    expect(didDoc.assertionMethod).toBeUndefined()
-
-    expect(didDoc.capabilityDelegation).toBeUndefined()
-
-    expect(didDoc.verificationMethod).toHaveLength(1)
-    expect(didDoc.verificationMethod).toContainEqual({
-      id: `${didDetails.did}#authentication`,
-      controller: didDetails.did,
-      type: 'Sr25519VerificationKey2020',
-      publicKeyBase58: base58Encode(authenticationDidKeyDetails.publicKey),
-    })
-
-    expect(didDoc.service).toBeUndefined()
-  })
-
-  it('does not export a DID Document with an unsupported format', () => {
-    const didCreationDetails: LightDidDetailsCreationOpts = {
-      authenticationKey: authenticationDidKeyDetails,
-    }
-    const didDetails = new LightDidDetails(didCreationDetails)
-    expect(() => didDetails.toDidDocument('text/html')).toThrow()
   })
 })
