+++ packages/kilt-did/src/index.ts
@@ -13,3 +13,4 @@ export * as DidUtils from './Did.utils'
 export * from './DemoKeystore/DemoKeystore'
 export * from './DidDetails'
 export * from './DidResolver/DefaultResolver'
+export * from './DidDocumentExporter/DidDocumentExporter'
