+++ packages/types/src/index.ts
@@ -26,3 +26,4 @@ export * from './Blockchain'
 export * from './DidDetails'
 export * from './Keystore'
 export * from './DidResolver'
+export * from './DidDocumentExporter'
